fix: redesigned map portal
